resolve compilation issues

+++ atoma-vllm/src/sequence.rs
@@ -5,7 +5,7 @@ use std::{
     time::{Duration, Instant},
 };
 
-use candle::Tensor;
+use candle_core::Tensor;
 use thiserror::Error;
 use tracing::{error, info, info_span, instrument, Span};
 
@@ -1018,7 +1018,7 @@ pub struct SequenceGroupMetadata {
     /// Stopping criteria parameters
     pub stopping_criteria_params: StoppingCriteriaParameters,
     /// Block tables
-    pub block_tables: HashMap<u64, Vec<i32>>,
+    pub block_tables: HashMap<u64, Vec<i64>>,
     /// Do sample (bool)
     pub do_sample: bool,
     /// Token chunk size
@@ -1038,7 +1038,7 @@ impl SequenceGroupMetadata {
         sequence_data: HashMap<u64, SequenceData>,
         next_token_chooser_params: NextTokenChooserParameters,
         stopping_criteria_params: StoppingCriteriaParameters,
-        block_tables: HashMap<u64, Vec<i32>>,
+        block_tables: HashMap<u64, Vec<i64>>,
         do_sample: bool,
         token_chunk_size: Option<usize>,
         state: SequenceGroupState,
@@ -1095,7 +1095,7 @@ pub struct SequenceOutput {
 }
 
 /// `SequenceGroupMetrics` - Metrics for a sequence group token generation
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Default)]
 pub struct SequenceGroupMetrics {
     /// Time taken to generate the batched output
     pub time_to_generate: Option<f32>,
